pyproject.toml
[project]
name = "financial-freedom-api"
version = "0.1.0"
description = "Litestar API for the Financial Freedom budgeting app."
requires-python = ">=3.11"
dependencies = [
  "litestar[standard]>=2.18.0",
  "psycopg[binary]>=3.2.0",
  "python-dotenv>=1.0.1",
  "httpx>=0.28.0",
]

[project.optional-dependencies]
dev = [
  "pytest>=8.4.0",
]

[build-system]
requires = ["setuptools>=68", "wheel"]
build-backend = "setuptools.build_meta"

[tool.setuptools]
package-dir = {"" = "src"}

[tool.setuptools.packages.find]
where = ["src"]
